% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisMetaboData.R
\docType{class}
\name{AnalysisMetaboData-class}
\alias{AnalysisMetaboData-class}
\title{Class of data for statistical analysis}
\description{
Class of data for statistical analysis
}
\section{Slots}{

\describe{
\item{\code{data}}{data frame that contains animal, specified observations, absolute time (my time), floating point day (RelDay), day activity (Sun), sine function based on day activity, squared floating point day (SqRelDay)}

\item{\code{animal}}{name of animal column}

\item{\code{actSwitchHour}}{hour at which light is switch on}
}}

