% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisMetaboData.R
\name{initialize,AnalysisMetaboData-method}
\alias{initialize,AnalysisMetaboData-method}
\title{Constructor of AnalysisMetaboData}
\usage{
\S4method{initialize}{AnalysisMetaboData}(
  .Object,
  rawData,
  date = "Date",
  time = "Time",
  actSwitchHour = "6",
  animal = "Animal No.",
  obs = "VO2(3)",
  annotation = character(0),
  annotGroups = c("Group")
)
}
\arguments{
\item{rawData}{S4 object of RawMetaboData}

\item{date}{name of date column in raw data}

\item{time}{name of time column in raw data}

\item{actSwitchHour}{hour at which light is switched on}

\item{animal}{name of animal column in raw data}

\item{obs}{list of observation to be extracted in raw data}

\item{annotation}{data frame of annotation to be added to each animals, must contain the column Animal, Time, Date}

\item{annotGroups}{list of qualitative annotation column. Other columns will be treated as numeric with a time dependence evaluated by spline.}
}
\description{
Constructor of AnalysisMetaboData
}
