% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResDailyMeanStatMetabo.R
\name{initialize,ResDailyMeanStatMetabo-method}
\alias{initialize,ResDailyMeanStatMetabo-method}
\title{Constructor for ResStatMetabo, perform a mixed linear statistical test}
\usage{
\S4method{initialize}{ResDailyMeanStatMetabo}(
  .Object,
  anMetData,
  observation,
  norm = NULL,
  group,
  control = "control",
  hourWin = c(8, 20),
  meanAnimal = F,
  statLog = F,
  timWind = c()
)
}
\arguments{
\item{anMetData}{S4 object of AnalysisMetaboData}

\item{observation}{column of anMetData used for analysis}

\item{norm}{column of anMetData used for normalization (if NULL, no normalization is applied)}

\item{group}{column of anMetData used experimental annotation}

\item{control}{name of control in experimental annotation}

\item{hourWin}{window (in hours) for applying mean. If it is empty, night is chosen}

\item{meanAnimal}{true if the mean is applied onto the animal, otherwise only on time window}

\item{statLog}{if TRUE, log10 is applied to observation (after normalization) for statistical analysis}

\item{timWind}{time window (in days) on which model is applied}
}
\description{
Constructor for ResStatMetabo, perform a mixed linear statistical test
}
