% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResStatMetabo.R
\name{initialize,ResStatMetabo-method}
\alias{initialize,ResStatMetabo-method}
\title{Constructor for ResStatMetabo, perform a mixed linear statistical test}
\usage{
\S4method{initialize}{ResStatMetabo}(
  .Object,
  anMetData,
  observation,
  model = "quadratic",
  norm = NULL,
  group,
  control = "control",
  statLog = F,
  timWind = c()
)
}
\arguments{
\item{anMetData}{S4 object of AnalysisMetaboData}

\item{observation}{column of anMetData used for analysis}

\item{model}{type of time model (constant, linear, quadratic, linearReg)}

\item{norm}{column of anMetData used for normalization (if NULL, no normalization is applied)}

\item{group}{column of anMetData used experimental annotation}

\item{control}{name of control in experimental annotation}

\item{statLog}{if TRUE, log10 is applied to observation (after normalization) for statistical analysis}

\item{timWind}{time window (in days) on which model is applied}
}
\description{
Constructor for ResStatMetabo, perform a mixed linear statistical test
}
