% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResDailyMeanStatMetabo.R
\name{metaboDailyPlot,ResDailyMeanStatMetabo-method}
\alias{metaboDailyPlot,ResDailyMeanStatMetabo-method}
\title{Plot time dependant metabolic data}
\usage{
\S4method{metaboDailyPlot}{ResDailyMeanStatMetabo}(x, signif = T, pvalStar = T, mainTitle = "", cex.axis.lab = 1)
}
\arguments{
\item{x}{ResDailyMeanStatMetabo S4 object}

\item{signif}{true for significance pairwise annotation}

\item{pvalStar}{significant annotation with stars instead of p-value}

\item{cex.axis.lab}{cex of axis tick, label, title and pval}

\item{type}{type of plot: data, data.model or model}
}
\description{
Plot time dependant metabolic data
}
