% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResStatMetabo.R
\name{metaboPlot,ResStatMetabo-method}
\alias{metaboPlot,ResStatMetabo-method}
\title{Plot time dependant metabolic data}
\usage{
\S4method{metaboPlot}{ResStatMetabo}(
  x,
  type = "data",
  mainTitle = "",
  cex.axis.lab = 1,
  inner.title = "",
  inner.left.pos = NaN
)
}
\arguments{
\item{x}{ResStatMetabo S4 object}

\item{type}{type of plot: data, data.model or model}

\item{mainTitle}{title}

\item{inner.title}{title inside the plot, next to legend}

\item{cex.lab.axis}{cex of axis tick, label, title and pval}

\item{inner.let.pos}{xposition of inner title}
}
\description{
Plot time dependant metabolic data
}
