% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisMetaboData.R
\name{metaboRawPlot,AnalysisMetaboData-method}
\alias{metaboRawPlot,AnalysisMetaboData-method}
\title{Plot time dependant metabolic  raw data}
\usage{
\S4method{metaboRawPlot}{AnalysisMetaboData}(x, observation, type = "data", group = "Group")
}
\arguments{
\item{x}{AnalysisMetaboData S4 object}

\item{type}{type of plot: data, mean.sd}

\item{group}{Group for coloring and/or mean/sd}
}
\description{
Plot time dependant metabolic  raw data
}
