% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawMetaboData.R
\docType{class}
\name{RawMetaboData-class}
\alias{RawMetaboData-class}
\title{Class of raw data parsing}
\description{
Class of raw data parsing
}
\section{Slots}{

\describe{
\item{\code{data}}{the data frame of raw data}

\item{\code{header}}{the rest of information contained in raw data file}
}}

