% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResStatMetabo.R
\docType{class}
\name{ResStatMetabo-class}
\alias{ResStatMetabo-class}
\title{Class for results of statistical analysis}
\description{
Class for results of statistical analysis
}
\section{Slots}{

\describe{
\item{\code{data}}{data used for analysis}

\item{\code{observation}}{name of output extracted for AnalysisMetaboData}

\item{\code{norm}}{name of data in AnalysisMetaboData used for normalization}

\item{\code{group}}{name of data in AnalysisMetaboData used for experimental annotation}

\item{\code{model}}{type time model (constant, linear, quadratic)}

\item{\code{lmeRes}}{result of mixed linear model}
}}

