% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisMetaboData.R
\name{initialize,AnalysisMetaboData-method}
\alias{initialize,AnalysisMetaboData-method}
\title{Constructor of AnalysisMetaboData}
\usage{
\S4method{initialize}{AnalysisMetaboData}(
  .Object,
  rawData,
  date = "Date",
  time = "Time",
  animal = "Animal No.",
  obs = "VO2(3)",
  annotation = NULL,
  annotGroups = c("Group")
)
}
\arguments{
\item{rawData}{S4 object of RawMetaboData}

\item{date}{name of date column}

\item{time}{name of time column}

\item{animal}{name of animal column}

\item{obs}{list of observation to be extracted}

\item{annotation}{data frame of annotation to be added to each animals, must contains an Animal column}
}
\description{
Constructor of AnalysisMetaboData
}
