% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawMetaboData.R
\name{initialize,RawMetaboData-method}
\alias{initialize,RawMetaboData-method}
\title{Constructor of class RawMetaboData}
\usage{
\S4method{initialize}{RawMetaboData}(.Object, fileName)
}
\arguments{
\item{fileName}{.csv file containing raw data}
}
\description{
Constructor of class RawMetaboData
}
