% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResStatMetabo.R
\name{metaboPlot,ResStatMetabo-method}
\alias{metaboPlot,ResStatMetabo-method}
\title{Plot time dependant metabolic data}
\usage{
\S4method{metaboPlot}{ResStatMetabo}(x, type = "data", mainTitle = "")
}
\arguments{
\item{x}{ResStatMetabo S4 object}

\item{type}{type of plot: data, data.model or model}

\item{mainTitle}{title}
}
\description{
Plot time dependant metabolic data
}
