% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawMetaboData.R
\docType{class}
\name{RawMetaboData-class}
\alias{RawMetaboData-class}
\title{Class of raw data parsing}
\description{
Class of raw data parsing
}
\section{Slots}{

\describe{
\item{\code{Date}}{start recording date}

\item{\code{Batch}}{Batch ID of the recording}

\item{\code{Device}}{Device model}

\item{\code{header}}{the rest of information contained in raw data file}

\item{\code{data}}{the data frame of raw data, one line is a measure, it is sort by animal}
}}

