% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResDailyMeanStatMetabo.R
\docType{class}
\name{ResDailyMeanStatMetabo-class}
\alias{ResDailyMeanStatMetabo-class}
\title{Class for linear modeling of temporal mean}
\description{
Class for linear modeling of temporal mean
}
\section{Slots}{

\describe{
\item{\code{observation}}{name of output extracted for AnalysisMetaboData}

\item{\code{norm}}{name of data in AnalysisMetaboData used for normalization}

\item{\code{group}}{name of data in AnalysisMetaboData used for experimental annotation}

\item{\code{hourWin}}{window (in hours) for applying mean}

\item{\code{meanAnimal}}{true if the mean is applied onto the animal, otherwise only on time window}

\item{\code{lmRes}}{result of mixed linear model}

\item{\code{statLog}}{if TRUE, log10 is applied to observation (after normalization) for statistical analysis}

\item{\code{timWind}}{time window (in days) on which model is applied}
}}

