% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResStatMetabo.R
\docType{class}
\name{ResStatMetabo-class}
\alias{ResStatMetabo-class}
\title{Class for results of statistical analysis}
\description{
Class for results of statistical analysis
}
\section{Slots}{

\describe{
\item{\code{data}}{data used for analysis}

\item{\code{observation}}{name of output extracted for AnalysisMetaboData}

\item{\code{norm}}{name of data in AnalysisMetaboData used for normalization}

\item{\code{group}}{name of data in AnalysisMetaboData used for experimental annotation}

\item{\code{actSwitchHour}}{hour at which light is switch on}

\item{\code{model}}{type time model (constant, linear, quadratic, linearReg)}

\item{\code{lmeRes}}{result of mixed linear model}

\item{\code{statLog}}{if TRUE, log10 is applied to observation (after normalization) for statistical analysis}

\item{\code{timWind}}{time window (in days) on which model is applied}
}}

